Remove unused modules and switch all hooks to use guet.hooks
