+++ guet/__init__.py
@@ -16,4 +16,3 @@ limitations under the License.
 
 __version__ = '2.0.0'
 
-from guet.commit import PostCommitManager
