+++ guet/git/create_hook.py
@@ -1,23 +1,26 @@
 from enum import Enum
 from os import chmod, stat
-
-from guet.git._create_hook_post_commit_strategy import _PostCommitStrategy
-from guet.git._create_hook_pre_commit_strategy import _PreCommitStrategy
-from guet.git._create_hook_commit_msg_strategy import _CommitMsgStrategy
+from os.path import join
 from guet.git._hook_mode import HookMode
+
+HOOK_FILE_LINES = [
+    '#! /usr/bin/env python3\n',
+    'from guet.hooks import manage\n',
+    'import sys\n',
+    'manage(sys.argv[0])\n',
+]
 
 
 class Hooks(Enum):
-    PRE_COMMIT = _PreCommitStrategy
-    COMMIT_MSG = _CommitMsgStrategy
-    POST_COMMIT = _PostCommitStrategy
+    PRE_COMMIT = 'pre-commit'
+    COMMIT_MSG = 'commit-msg'
+    POST_COMMIT = 'post-commit'
 
 
 def create_hook(hook_folder_path: str, hook: Hooks, create_mode: HookMode):
-    hook_strategy = hook.value(hook_folder_path, create_mode)
-    final_hook_path = hook_strategy.final_hook_path()
+    final_hook_path = join(hook_folder_path, create_mode(hook.value))
     f = open(final_hook_path, 'w')
-    f.writelines(hook_strategy.file_lines())
+    f.writelines(HOOK_FILE_LINES)
     f.close()
 
     st = stat(final_hook_path)
